[Refactor] UserBook 도메인 Converter 분리
